refactor: create markdown util library
With parseMarkdownHtml and stripMarkdown functions, formerly mixins
