+++ packages/portal/src/components/landing/LandingHero.vue
@@ -40,7 +40,7 @@
 <script>
   import ImageWithAttribution from '@/components/image/ImageWithAttribution';
   import SmartLink from '@/components/generic/SmartLink';
-  import parseMarkdownHtmlMixin from '@/mixins/parseMarkdownHtml';
+  import { parseMarkdownHtml } from '@/utils/markdown.js';
 
   export default {
     name: 'LandingHero',
@@ -49,8 +49,6 @@
       ImageWithAttribution,
       SmartLink
     },
-
-    mixins: [parseMarkdownHtmlMixin],
 
     props: {
       /**
@@ -109,6 +107,10 @@
       isSVG() {
         return this.heroImage?.image?.contentType === 'image/svg+xml';
       }
+    },
+
+    methods: {
+      parseMarkdownHtml
     }
   };
 </script>
