+++ packages/portal/src/components/landing/LandingImageCardGroup.vue
@@ -35,15 +35,13 @@
 </template>
 
 <script>
-  import parseMarkdownHtmlMixin from '@/mixins/parseMarkdownHtml';
+  import { parseMarkdownHtml } from '@/utils/markdown.js';
 
   export default {
     name: 'LandingImageCardGroup',
     components: {
       LandingImageCard: () => import('@/components/landing/LandingImageCard')
     },
-
-    mixins: [parseMarkdownHtmlMixin],
 
     props: {
       /**
@@ -82,6 +80,10 @@
           'bg-color-highlight': this.backgroundImage?.profile?.background === 'highlight'
         }
       };
+    },
+
+    methods: {
+      parseMarkdownHtml
     }
   };
 </script>
