+++ packages/portal/src/components/story/StoryHero.vue
@@ -51,7 +51,6 @@
 </template>
 
 <script>
-  import parseMarkdownHtmlMixin from '@/mixins/parseMarkdownHtml';
   import ImageWithAttribution from '@/components/image/ImageWithAttribution';
   import { FULL_VIEWPORT_PRESETS } from '@/utils/contentful/imageCropPresets';
 
@@ -61,8 +60,6 @@
     components: {
       ImageWithAttribution
     },
-
-    mixins: [parseMarkdownHtmlMixin],
 
     props: {
       title: {
