--- packages/portal/src/mixins/parseMarkdownHtml.js
@@ -1,9 +1,0 @@
-import { marked } from 'marked';
-
-export default {
-  methods: {
-    parseMarkdownHtml(text) {
-      return text ? marked.parse(text) : text;
-    }
-  }
-};
